twentieth commit
At the moment, the most popular products and search section on the home page and parts of the single product page and front and security issues and preparation for deployment on the host are left. Maybe there are other things, but I don't remember now. This is the twentieth commit.